Replaced log viewer activity with Compose screen, removed unused code from log
