+++ WooCommerce/src/main/kotlin/com/woocommerce/android/util/RollingLogEntries.kt
@@ -2,11 +2,6 @@ package com.woocommerce.android.util
 
 import com.woocommerce.android.util.RollingLogEntries.LogEntry
 import com.woocommerce.android.util.WooLog.LogLevel
-import com.woocommerce.android.util.WooLog.LogLevel.d
-import com.woocommerce.android.util.WooLog.LogLevel.e
-import com.woocommerce.android.util.WooLog.LogLevel.i
-import com.woocommerce.android.util.WooLog.LogLevel.v
-import com.woocommerce.android.util.WooLog.LogLevel.w
 import com.woocommerce.android.util.WooLog.T
 import org.wordpress.android.util.DateTimeUtils
 import java.security.InvalidParameterException
@@ -27,29 +22,6 @@ class RollingLogEntries(private val limit: Int) : LinkedList<LogEntry>() {
             removeFirst()
         }
         return super.add(element)
-    }
-
-    /**
-     * Returns the log entries as an array of html-formatted strings - this enables us to display
-     * a formatted log in [com.woocommerce.android.support.WooLogViewerActivity]
-     */
-    fun toHtmlList(isDarkTheme: Boolean): List<String> {
-        fun LogEntry.getColor(): String {
-            return if (isDarkTheme) {
-                "white"
-            } else when (level) {
-                v -> "grey"
-                d -> "teal"
-                i -> "black"
-                w -> "purple"
-                e -> "red"
-            }
-        }
-
-        // work with a copy of the log entries in case they're modified while traversing them
-        return toList().map { entry ->
-            "<font color='${entry.getColor()}'>$entry</font>"
-        }
     }
 
     /**
